refactorized component names
