+++ client/components/icons/IconLoginAsUserComponent.vue
