+++ client/components/icons/IconSearchComponent.vue
